Move grdcut to gridops.py
