+++ pygmt/__init__.py
@@ -18,7 +18,8 @@ from .filtering import blockmedian
 from .gridding import surface
 from .sampling import grdtrack
 from .mathops import makecpt
-from .modules import config, info, grdinfo, grdcut, which
+from .modules import config, info, grdinfo, which
+from .gridops import grdcut
 from . import datasets
